test: rework minio server for mock tests
* All tests use the same server
* An already running server is used if it exists
