--- src/test/java/org/janelia/saalfeldlab/n5/s3/mock/RunnerWithMinioServer.java
@@ -1,191 +1,0 @@
-package org.janelia.saalfeldlab.n5.s3.mock;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.internal.runners.statements.RunAfters;
-import org.junit.internal.runners.statements.RunBefores;
-import org.junit.runner.Description;
-import org.junit.runner.notification.RunNotifier;
-import org.junit.runners.BlockJUnit4ClassRunner;
-import org.junit.runners.model.FrameworkField;
-import org.junit.runners.model.FrameworkMethod;
-import org.junit.runners.model.Statement;
-import org.junit.runners.model.TestClass;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-public class RunnerWithMinioServer extends BlockJUnit4ClassRunner {
-
-	private Process process;
-
-	private final StringBuilder perTestHttpOut = new StringBuilder();
-
-	public Path minioServerDirectory;
-
-	public final URI minioUri = URI.create("http://localhost:9000/");
-
-	public RunnerWithMinioServer(Class<?> klass) throws Exception {
-
-		super(klass);
-	}
-
-	private static Path createTmpServerDirectory() throws IOException {
-
-		/* deleteOnExit doesn't work on temporary files, so delete it manually and recreate explicitly...*/
-		final Path tempDirectory = Files.createTempDirectory("n5-minio-test-server-");
-		tempDirectory.toFile().delete();
-		tempDirectory.toFile().mkdirs();
-		tempDirectory.toFile().deleteOnExit();
-		return tempDirectory;
-	}
-
-	@Override protected Object createTest() throws Exception {
-
-		final Object test = super.createTest();
-		for (FrameworkField field : getTestClass().getAnnotatedFields(Deprecated.class)) {
-			if (field.getType().isAssignableFrom(Path.class)) {
-				field.getField().set(test, minioServerDirectory);
-			} else if (field.getType().isAssignableFrom(URI.class)) {
-				field.getField().set(test, minioUri);
-			}
-		}
-		return test;
-	}
-
-	@Override protected void runChild(FrameworkMethod method, RunNotifier notifier) {
-
-		if (!process.isAlive()) {
-			logMinioOutput();
-			return;
-		}
-
-		Description description = describeChild(method);
-		if (isIgnored(method)) {
-			notifier.fireTestIgnored(description);
-		} else {
-			Statement statement = new Statement() {
-
-				@Override
-				public void evaluate() throws Throwable {
-
-					try {
-						methodBlock(method).evaluate();
-					} catch (Exception e) {
-						if (!process.isAlive())
-							logMinioOutput();
-						throw e;
-					} finally {
-						perTestHttpOut.setLength(0);
-					}
-				}
-			};
-			runLeaf(statement, description, notifier);
-		}
-
-	}
-
-	private void logMinioOutput() {
-
-		if (perTestHttpOut.length() > 0) {
-			perTestHttpOut.insert(0, "Last Minio Server Output.\n");
-			perTestHttpOut.insert(0, "Minio Server is not alive.\n");
-			System.err.println(perTestHttpOut);
-		}
-	}
-
-	@Before
-	public void startMinioServer() throws Exception {
-		
-		if( isMinioServerRunning() ) {
-			return;
-		}
-
-		// TODO if the server fails to start for some reason
-		// e.g. minio server not installed
-		// probably should not fail, but report "mock test skipped" or something
-		minioServerDirectory = createTmpServerDirectory();
-		ProcessBuilder processBuilder = new ProcessBuilder("minio", "server", ".");
-		processBuilder.directory(minioServerDirectory.toFile());
-		processBuilder.redirectErrorStream(true);
-		process = processBuilder.start();
-		waitForReady();
-		/* give the server some time to finish startup */
-		final Thread clearStdout = new Thread(() -> {
-			try (BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()))) {
-				String line;
-				while ((line = reader.readLine()) != null) {
-					perTestHttpOut.append(line).append("\n");
-				}
-			} catch (IOException e) {
-			}
-		});
-		clearStdout.setDaemon(true);
-		clearStdout.start();
-	}
-
-	private boolean isMinioServerRunning() {
-
-		try {
-			minioUri.toURL().openConnection().connect();
-			return true;
-		} catch (IOException e) {
-		}
-		return false;
-	}
-
-	private void waitForReady() throws IOException, InterruptedException {
-
-		final Thread waitForConnect = new Thread(() -> {
-			while (true) {
-				try {
-					minioUri.toURL().openConnection().connect();
-					return;
-				} catch (Exception e) {
-					 try {
-						Thread.sleep(100);
-					} catch (InterruptedException e1) {
-					}
-				}
-			}
-		});
-		waitForConnect.start();
-		waitForConnect.join(10_000);
-	}
-
-	@After
-	public void stopMinioServer() {
-
-		// we didn't start the server ourselves
-		if (process == null)
-			return;
-
-		process.destroy();
-		try {
-			process.waitFor(1, TimeUnit.SECONDS);
-		} catch (InterruptedException e) {
-			process.destroyForcibly();
-		}
-	}
-
-	@Override protected Statement withBeforeClasses(Statement statement) {
-
-		final Statement testClassBefore = super.withBeforeClasses(statement);
-		final List<FrameworkMethod> beforeTestClass = new TestClass(RunnerWithMinioServer.class).getAnnotatedMethods(Before.class);
-		return new RunBefores(testClassBefore, beforeTestClass, this);
-	}
-
-	@Override protected Statement withAfterClasses(Statement statement) {
-
-		final List<FrameworkMethod> afterTestClass = new TestClass(RunnerWithMinioServer.class).getAnnotatedMethods(After.class);
-		final RunAfters runnerAfterClass = new RunAfters(statement, afterTestClass, this);
-		return super.withAfterClasses(runnerAfterClass);
-	}
-
-}
